refactor setting router code
splitted up one huge router settings file into the smaller files
representing the actual page structure
